+++ sdk/resourcemanager/azure-resourcemanager-storage/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.2.0-beta.1 (Unreleased)
 
 - Updated `api-version` to `2021-01-01`
-- Return type of `Identity.type()` changed to `IdentityType`
+- Return type of `Identity.type()` changed from `String` to `IdentityType`
 
 ## 2.1.0 (2020-11-24)